refactor(java): Create a DefaultFieldsFilter to implement FieldsFilter.
