+++ alert-common/src/main/java/com/synopsys/integration/alert/common/security/authorization/AuthorizationManager.java
@@ -60,13 +60,13 @@ public class AuthorizationManager {
     public final boolean hasPermissions(final String permissionKey) {
         Collection<String> roleNames = getCurrentUserRoleNames();
         return roleNames.stream()
-                   .anyMatch(name -> permissionCache.containsKey(name) && !permissionCache.get(name).hasPermissions(permissionKey));
+                   .anyMatch(name -> permissionCache.containsKey(name) && permissionCache.get(name).hasPermissions(permissionKey));
     }
 
     public final boolean isReadOnly(final String permissionKey) {
         Collection<String> roleNames = getCurrentUserRoleNames();
         return roleNames.stream()
-                   .allMatch(name -> permissionCache.containsKey(name) && !permissionCache.get(name).isReadOnly(permissionKey));
+                   .allMatch(name -> permissionCache.containsKey(name) && permissionCache.get(name).isReadOnly(permissionKey));
     }
 
     public final boolean hasCreatePermission(final PermissionKeys permissionKey) {
