+++ src/main/java/com/synopsys/integration/alert/channel/email/EmailDistributionFieldsFilter.java
@@ -20,7 +20,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package com.synopsys.integration.alert.component.scheduling;
+package com.synopsys.integration.alert.channel.email;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -28,14 +28,11 @@ import org.springframework.stereotype.Component;
 import com.synopsys.integration.alert.common.descriptor.config.filter.DefaultFieldsFilter;
 import com.synopsys.integration.alert.common.enumeration.ConfigContextEnum;
 import com.synopsys.integration.alert.common.security.authorization.AuthorizationManager;
-import com.synopsys.integration.alert.component.scheduling.descriptor.SchedulingDescriptor;
 
 @Component
-public class SchedulingFieldsFilter extends DefaultFieldsFilter {
-
+public class EmailDistributionFieldsFilter extends DefaultFieldsFilter {
     @Autowired
-    public SchedulingFieldsFilter(final AuthorizationManager authorizationManager) {
-        super(SchedulingDescriptor.SCHEDULING_COMPONENT, ConfigContextEnum.GLOBAL, authorizationManager);
+    public EmailDistributionFieldsFilter(final AuthorizationManager authorizationManager) {
+        super(EmailChannel.COMPONENT_NAME, ConfigContextEnum.DISTRIBUTION, authorizationManager);
     }
 }
-
